Reading Data

Base.read — Function.

read(f::FortranFile [, spec [, spec [, ...]]])

Read data from a FortranFile. Like the READ statement in Fortran, this reads a completely record, regardless of how man specs are given. Each spec can be:

a DataType for scalar values; e.g. Int32, Float64, FString{10}

a tuple of DataType and one or more integers, for reading arrays of the given size; e.g. (Int32,4,2) reads an Array{Int32}(4,2)

a tuple of DataType and a tuple of integers, as an alternative way of reading arrays; e.g. (Int32,(4,2)) does the same as the previous one

an array, for reading into pre-allocated arrays; DataType and size of the array are implied through its Julia type.

Return value:

if no spec is given: nothing (the record is skipped over)

if one spec is given: the scalar or array requested

if more specs are given: a tuple of the scalars and arrays requested, in order

source

Examples

The following examples show how to write Julia code that corresponds to certain Fortran READ statements. The Julia code assumes that f refers to an opened FortranFile, while the Fortran code assumes that lun refers to a logical unit number for a connected file.

Reading a single scalar

x = read(f, Float64)

corresponds to

real(kind=real64)::x
read(lun) x

Reading a 1D array

vector = read(f, (Float64,10))       # read into a new array
vector = zeros(10); read(f, vector)  # read into pre-existing array

corresponds to (Modern Fortran style)

real(kind=real64),dimension(10)::vector
read(lun) vector

and to (Fortran77 style)

integer::i
real(kind=real64),dimension(10)::vector
read(lun) (vector(i), i=1,10)

Reading a 2D array

matrix = read(f, (Float64,10,10))      # read into a new array
matrix = read(f, (Float64,(10,10)))    # alternative syntax
matrix = zeros(10,10); read(f, matrix) # read into existing array

corresponds to

real(kind=real64),dimension(10,10)::matrix
read(lun) matrix

Reading a character string

fstr = read(f, FString{20})

corresponds to

character(len=20)::fstr
read(lun) fstr

Reading a record with multiple data

i, strings, zmatrix = read(f, Int32, (Fstring{20},10), (Complex128,10,10))

corresponds to

integer(kind=int32)::i
character(len=20),dimension(10)::strings
complex(kind=real64),dimension(10,10)::zmatrix
read(lun) i,strings,matrix

Skipping over a record

read(f)

corresponds to

read(lun)